requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
batchrenorm==0.1.0
bcrypt==3.2.0
bleach==4.0.0
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.0.1
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
Distance==0.1.3
einops==0.4.1
entrypoints==0.3
fabric==2.6.0
filelock==3.4.0
fsspec==2021.7.0
future==0.18.2
g2p-en==2.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.36.1
huggingface-hub==0.2.1
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.4
inflect==5.3.0
invoke==1.6.0
ipykernel==5.3.4
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonargparse==3.18.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
Montreal-Forced-Aligner==2.0.0b3
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.7
notebook==6.4.3
numba==0.53.1
numexpr==2.7.3
numpy==1.20.3
oauthlib==3.1.1
olefile==0.46
packaging==21.0
pandas==1.3.5
pandocfilters==1.4.3
paramiko==2.8.0
parso==0.8.2
pathlib2==2.3.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.2.4
plotly==5.14.1
pooch==1.4.0
praatio==5.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.17
protobuf==3.17.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.1
Pygments==2.10.0
PyJWT==2.1.0
PyNaCl==1.4.0
pynini==2.1.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypinyin==0.44.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2021.3
pyu2f==0.1.5
pyworld==0.3.0
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.0
QtPy==1.10.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==0.24.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.16.0
SoundFile==0.10.3.post1
tenacity==8.2.2
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
test-tube==0.7.5
testpath==0.5.0
thop==0.1.1.post2209072238
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==0.5.0
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
transformers==4.14.1
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.5.0
